requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
motor==3.7.0
pymongo==4.12.1
openai==1.78.1
pinecone==6.0.2
crawl4ai==0.6.3
apscheduler==3.11.0
duckduckgo-search==7.3.2
python-multipart==0.0.20
google-generativeai==0.8.4
httpx==0.28.1
